feat(sharing): move avatar-sharing to herobox, use addthis instead of custom solution. refactor from pages/* to export/*
