+++ src/controllers/dataexport.js
@@ -8,6 +8,12 @@ var js2xmlparser = require("js2xmlparser");
 var pd = require('pretty-data').pd;
 var User = require('../models/user').model;
 
+// Avatar screenshot/static-page includes
+var Pageres = require('pageres'); //https://github.com/sindresorhus/pageres
+var AWS = require('aws-sdk');
+AWS.config.update({accessKeyId: nconf.get("S3").accessKeyId, secretAccessKey: nconf.get("S3").secretAccessKey});
+var s3Stream = require('s3-upload-stream')(new AWS.S3()); //https://github.com/nathanpeck/s3-upload-stream
+var bucket = 'habitrpg-dev';
 
 /*
   ------------------------------------------------------------------------
@@ -82,4 +88,43 @@ expressres.jsonstring = function(obj, headers, status) {
   this.header('Content-Disposition', 'attachment');
   body = pd.json(JSON.stringify(obj));
   return this.send(body, headers, status);
+};
+
+/*
+ ------------------------------------------------------------------------
+ Static page and image screenshot of avatar
+ ------------------------------------------------------------------------
+ */
+
+
+dataexport.avatarPage = function(req, res) {
+  User.findById(req.params.uuid).select('stats profile items achievements preferences backer contributor').exec(function(err, user){
+    res.render('avatar-static', {
+      title: user.profile.name,
+      env: _.defaults({user:user},res.locals.habitrpg)
+    });
+  })
+};
+
+dataexport.avatarImage = function(req, res, next) {
+  var filename = 'avatar-'+req.params.uuid+'.png';
+  new Pageres()//{delay:1}
+    .src(nconf.get('BASE_URL')+'/export/avatar-'+req.params.uuid+'.html', ['140x147'], {crop: true, filename: filename.replace('.png','')})
+    .run(function (err, file) {
+      if (err) return next(err);
+      var upload = s3Stream.upload({
+        Bucket: bucket,
+        Key: filename,
+        ACL: "public-read",
+        StorageClass: "REDUCED_REDUNDANCY",
+        ContentType: "binary/octet-stream"
+      });
+      upload.on('error', function (err) {
+        next(err);
+      });
+      upload.on('uploaded', function (details) {
+        res.redirect(details.Location);
+      });
+      file[0].pipe(upload);
+    });
 };
