+++ src/routes/pages.js
@@ -28,42 +28,6 @@ _.each(pages, function(name){
   });
 })
 
-var mongoose = require('mongoose');
-router.get('/static/avatar-:uuid.html', i18n.getUserLanguage, middleware.locals, function(req, res) {
-  mongoose.model('User').findById(req.params.uuid).select('stats profile items achievements preferences backer contributor').exec(function(err, user){
-    res.render('avatar-static', {
-      title: user.profile.name,
-      env: _.defaults({user:user},res.locals.habitrpg)
-    });
-  })
-});
-var Pageres = require('pageres'); //https://github.com/sindresorhus/pageres
-var AWS = require('aws-sdk');
-AWS.config.update({accessKeyId: nconf.get("S3").accessKeyId, secretAccessKey: nconf.get("S3").secretAccessKey});
-var s3Stream = require('s3-upload-stream')(new AWS.S3()); //https://github.com/nathanpeck/s3-upload-stream
-var bucket = 'habitrpg-dev';
-
-router.get('/static/avatar-:uuid.png', i18n.getUserLanguage, middleware.locals, function(req, res, next) {
-  var filename = 'avatar-'+req.params.uuid+'.png';
-  new Pageres()//{delay:1}
-    .src(nconf.get('BASE_URL')+'/static/avatar-'+req.params.uuid+'.html', ['140x147'], {crop: true, filename: filename.replace('.png','')})
-    .run(function (err, file) {
-      if (err) return next(err);
-      var upload = s3Stream.upload({
-        Bucket: bucket,
-        Key: filename,
-        ACL: "public-read",
-        StorageClass: "REDUCED_REDUNDANCY",
-        ContentType: "binary/octet-stream"
-      });
-      file[0].pipe(upload);
-
-      upload.on('uploaded', function (details) {
-        res.redirect(details.Location);
-      });
-    });
-});
-
 // --------- Redirects --------
 
 router.get('/static/extensions', function(req, res) {
